Automação finalizada
